Move Avro/Protobuf functions to sql/api
