+++ sql/api/src/main/scala/org/apache/spark/sql/avro/functions.scala
@@ -21,8 +21,7 @@ import scala.jdk.CollectionConverters._
 
 import org.apache.spark.annotation.Experimental
 import org.apache.spark.sql.Column
-import org.apache.spark.sql.internal.ExpressionUtils.{column, expression}
-
+import org.apache.spark.sql.functions.lit
 
 // scalastyle:off: object.name
 object functions {
@@ -33,28 +32,30 @@ object functions {
    * schema must match the read data, otherwise the behavior is undefined: it may fail or return
    * arbitrary result.
    *
-   * @param data the binary column.
-   * @param jsonFormatSchema the avro schema in JSON string format.
+   * @param data
+   *   the binary column.
+   * @param jsonFormatSchema
+   *   the avro schema in JSON string format.
    *
    * @since 3.0.0
    */
   @Experimental
-  def from_avro(
-      data: Column,
-      jsonFormatSchema: String): Column = {
-    AvroDataToCatalyst(data, jsonFormatSchema, Map.empty)
+  def from_avro(data: Column, jsonFormatSchema: String): Column = {
+    Column.fn("from_avro", data, lit(jsonFormatSchema))
   }
 
   /**
-   * Converts a binary column of Avro format into its corresponding catalyst value.
-   * The specified schema must match actual schema of the read data, otherwise the behavior
-   * is undefined: it may fail or return arbitrary result.
-   * To deserialize the data with a compatible and evolved schema, the expected Avro schema can be
-   * set via the option avroSchema.
+   * Converts a binary column of Avro format into its corresponding catalyst value. The specified
+   * schema must match actual schema of the read data, otherwise the behavior is undefined: it may
+   * fail or return arbitrary result. To deserialize the data with a compatible and evolved
+   * schema, the expected Avro schema can be set via the option avroSchema.
    *
-   * @param data the binary column.
-   * @param jsonFormatSchema the avro schema in JSON string format.
-   * @param options options to control how the Avro record is parsed.
+   * @param data
+   *   the binary column.
+   * @param jsonFormatSchema
+   *   the avro schema in JSON string format.
+   * @param options
+   *   options to control how the Avro record is parsed.
    *
    * @since 3.0.0
    */
@@ -63,31 +64,34 @@ object functions {
       data: Column,
       jsonFormatSchema: String,
       options: java.util.Map[String, String]): Column = {
-    AvroDataToCatalyst(data, jsonFormatSchema, options.asScala.toMap)
+    Column.fnWithOptions("from_avro", options.asScala.iterator, data, lit(jsonFormatSchema))
   }
 
   /**
    * Converts a column into binary of avro format.
    *
-   * @param data the data column.
+   * @param data
+   *   the data column.
    *
    * @since 3.0.0
    */
   @Experimental
   def to_avro(data: Column): Column = {
-    CatalystDataToAvro(data, None)
+    Column.fn("to_avro", data)
   }
 
   /**
    * Converts a column into binary of avro format.
    *
-   * @param data the data column.
-   * @param jsonFormatSchema user-specified output avro schema in JSON string format.
+   * @param data
+   *   the data column.
+   * @param jsonFormatSchema
+   *   user-specified output avro schema in JSON string format.
    *
    * @since 3.0.0
    */
   @Experimental
   def to_avro(data: Column, jsonFormatSchema: String): Column = {
-    CatalystDataToAvro(data, Some(jsonFormatSchema))
+    Column.fn("to_avro", data, lit(jsonFormatSchema))
   }
 }
